+++ src/promptflow/tests/executor/e2etests/test_image.py
@@ -5,14 +5,14 @@ from tempfile import mkdtemp
 import pytest
 
 from promptflow._utils.multimedia_utils import MIME_PATTERN, _create_image_from_file, is_multimedia_dict
-from promptflow.batch import BatchEngine
+from promptflow.batch._batch_engine import OUTPUT_FILE_NAME, BatchEngine
+from promptflow.batch._result import BatchResult
 from promptflow.contracts.multimedia import Image
 from promptflow.contracts.run_info import FlowRunInfo, RunInfo, Status
 from promptflow.executor import FlowExecutor
-from promptflow.executor._result import BatchResult, LineResult
 from promptflow.storage._run_storage import DefaultRunStorage
 
-from ..utils import get_flow_folder, get_yaml_file, is_image_file, is_jsonl_file
+from ..utils import get_flow_folder, get_yaml_file, is_image_file, is_jsonl_file, load_jsonl
 
 SIMPLE_IMAGE_FLOW = "python_tool_with_simple_image"
 SAMPLE_IMAGE_FLOW_WITH_DEFAULT = "python_tool_with_simple_image_with_default"
@@ -86,34 +86,32 @@ def get_test_cases_for_node_run():
     ]
 
 
-def assert_contain_image_reference(value):
-    if isinstance(value, FlowRunInfo) or isinstance(value, RunInfo):
-        assert_contain_image_reference(value.api_calls)
-        assert_contain_image_reference(value.inputs)
-        assert_contain_image_reference(value.output)
+def contain_image_reference(value):
+    if isinstance(value, (FlowRunInfo, RunInfo)):
+        assert contain_image_reference(value.api_calls)
+        assert contain_image_reference(value.inputs)
+        assert contain_image_reference(value.output)
+        return True
     assert not isinstance(value, Image)
     if isinstance(value, list):
-        for item in value:
-            assert_contain_image_reference(item)
-    elif isinstance(value, dict):
+        return any(contain_image_reference(item) for item in value)
+    if isinstance(value, dict):
         if is_multimedia_dict(value):
             v = list(value.values())[0]
             assert isinstance(v, str)
-        else:
-            for _, v in value.items():
-                assert_contain_image_reference(v)
+            return True
+        return any(contain_image_reference(v) for v in value.values())
+    return False
 
 
-def assert_contain_image_object(value):
+def contain_image_object(value):
     if isinstance(value, list):
-        for item in value:
-            assert_contain_image_object(item)
+        return any(contain_image_object(item) for item in value)
     elif isinstance(value, dict):
         assert not is_multimedia_dict(value)
-        for _, v in value.items():
-            assert_contain_image_object(v)
+        return any(contain_image_object(v) for v in value.values())
     else:
-        assert isinstance(value, Image)
+        return isinstance(value, Image)
 
 
 @pytest.mark.usefixtures("dev_connections")
@@ -132,12 +130,14 @@ class TestExecutorWithImage:
         executor = FlowExecutor.create(get_yaml_file(flow_folder), dev_connections, storage=storage)
         flow_result = executor.exec_line(inputs)
         assert isinstance(flow_result.output, dict)
-        assert_contain_image_object(flow_result.output)
+        assert contain_image_object(flow_result.output)
+        # Assert output also contains plain text.
+        assert any(isinstance(v, str) for v in flow_result.output)
         assert flow_result.run_info.status == Status.Completed
-        assert_contain_image_reference(flow_result.run_info)
+        assert contain_image_reference(flow_result.run_info)
         for _, node_run_info in flow_result.node_run_infos.items():
             assert node_run_info.status == Status.Completed
-            assert_contain_image_reference(node_run_info)
+            assert contain_image_reference(node_run_info)
 
     @pytest.mark.parametrize(
         "flow_folder, node_name, flow_inputs, dependency_nodes_outputs", get_test_cases_for_node_run()
@@ -157,7 +157,7 @@ class TestExecutorWithImage:
             raise_ex=True,
         )
         assert run_info.status == Status.Completed
-        assert_contain_image_reference(run_info)
+        assert contain_image_reference(run_info)
 
     @pytest.mark.parametrize(
         "flow_folder, node_name, flow_inputs, dependency_nodes_outputs",
@@ -191,7 +191,7 @@ class TestExecutorWithImage:
             raise_ex=True,
         )
         assert run_info.status == Status.Completed
-        assert_contain_image_reference(run_info)
+        assert contain_image_reference(run_info)
 
     @pytest.mark.parametrize(
         "flow_folder, input_dirs, inputs_mapping, output_key, expected_outputs_number, has_aggregation_node",
@@ -257,24 +257,17 @@ class TestExecutorWithImage:
         )
 
         assert isinstance(batch_result, BatchResult)
+        assert batch_result.completed_lines == expected_outputs_number
         assert all(is_jsonl_file(output_file) or is_image_file(output_file) for output_file in output_dir.iterdir())
 
-        assert len(batch_result.outputs) == expected_outputs_number
-        for i, output in enumerate(batch_result.outputs):
+        outputs = load_jsonl(output_dir / OUTPUT_FILE_NAME)
+        assert len(outputs) == expected_outputs_number
+        for i, output in enumerate(outputs):
             assert isinstance(output, dict)
             assert "line_number" in output, f"line_number is not in {i}th output {output}"
             assert output["line_number"] == i, f"line_number is not correct in {i}th output {output}"
             result = output[output_key][0] if isinstance(output[output_key], list) else output[output_key]
             assert all(MIME_PATTERN.search(key) for key in result), f"image is not in {i}th output {output}"
-
-        for i, line_result in enumerate(batch_result.line_results):
-            assert isinstance(line_result, LineResult)
-            assert line_result.run_info.status == Status.Completed, f"{i}th line got {line_result.run_info.status}"
-
-        if has_aggregation_node:
-            for _, node_run_info in batch_result.aggr_results.node_run_infos.items():
-                assert node_run_info.status == Status.Completed
-                assert_contain_image_reference(node_run_info)
 
     @pytest.mark.parametrize(
         "flow_folder, inputs",
@@ -292,4 +285,4 @@ class TestExecutorWithImage:
         aggregation_results = executor.exec_aggregation(flow_inputs, aggregation_inputs=aggregation_inputs)
         for _, node_run_info in aggregation_results.node_run_infos.items():
             assert node_run_info.status == Status.Completed
-            assert_contain_image_reference(node_run_info)
+            assert contain_image_reference(node_run_info)