+++ src/promptflow/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Display node status in run visualize page graph view.
 - Add support for image input and output in prompt flow.
 - [SDK/CLI] SDK/CLI will collect telemetry by default, user can use `pf config set telemetry.enabled=false` to opt out.
+- Add `raise_on_error` for stream run API, by default we raise for failed run.
 
 ### Bugs Fixed